# application/services/custom_services/onedata/src/models/entitymodels/twodata.ts
export interface twodata 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   email: { type: String }
}